アクセス制限
what
投稿者のみ編集削除閲覧できるようにした
why
勝手に書き換えれないようにするため

+++ resources/views/items/show.blade.php
@@ -7,10 +7,14 @@
     </div>
     <div>
         <a href={{ route('items.index') }}>一覧に戻る</a>
-        | <a href={{ route('items.edit', ['id' =>  $item->id]) }}>編集</a>
-        <p></p>
-        {{ Form::open(['method' => 'delete', 'route' => ['items.destroy', $item->id]]) }}
-            {{ Form::submit('削除') }}
-        {{ Form::close() }}
+        @auth
+          @if ($item->user_id === $login_user_id)
+            | <a href={{ route('items.edit', ['id' =>  $item->id]) }}>編集</a>
+            <p></p>
+            {{ Form::open(['method' => 'delete', 'route' => ['items.destroy', $item->id]]) }}
+                {{ Form::submit('削除') }}
+            {{ Form::close() }}
+          @endif
+        @endauth
     </div>
 @endsection
